atividade8/producer/requirements.txt
kafka-python==2.0.2
pandas==2.2.2
python-dotenv==1.0.1